Remove BaseSearcgIT
